fix: revert SSL rename
